Use `react-hook-form` package for validations
