+++ packages/application/package.json
@@ -15,6 +15,7 @@
         "lodash": "^4.17.15",
         "react": "^16.13.1",
         "react-dom": "^16.13.1",
+        "react-hook-form": "^5.7.2",
         "react-router-dom": "^5.2.0",
         "react-scripts": "3.4.1",
         "typescript": "~3.7.2"
