+++ packages/website/src/index.tsx
@@ -8,11 +8,11 @@ import { store } from "./store/configure";
 import * as serviceWorker from "./serviceWorker";
 
 ReactDOM.render(
-  <>
+  <React.StrictMode>
     <Provider store={store}>
       <App />
     </Provider>
-  </>,
+  </React.StrictMode>,
   document.getElementById("root")
 );
