+++ packages/website/src/layout/App/index.tsx
@@ -5,7 +5,6 @@ import { BrowserRouter as Router, Switch, Route } from "react-router-dom";
 import NotFound from "../../routes/NotFound";
 import HomePage from "../../routes/Homepage";
 import ReefRoutes from "../../routes/ReefRoutes";
-import Form from "../../form";
 import theme from "./theme";
 import "./App.css";
 
@@ -17,7 +16,6 @@ function App() {
           <Switch>
             <Route exact path="/" component={HomePage} />
             <Route path="/reefs" component={ReefRoutes} />
-            <Route path="/form" component={Form} />
             <Route default component={NotFound} />
           </Switch>
         </div>
